query: report float64 argument fields as float64, not float32
FieldEditor.ObjectType mapped reflect.Float64 to astral.Float32, so a
float64 argument advertised half its width to every caller reading .spec.
The value was never wrong — UnmarshalText parses through ParseFloat(64) and
SetFloat — but a caller sizing or validating its input from the published
schema saw 32-bit precision.

Map it to astral.Float64, which exists and reports "float64".

Blast radius, which the defect report left unmeasured: no op-argument struct
in astral-go or astrald carries a float field today. astrald's only float
struct is LinkPressureConfig (mod/nodes/src/pressure.go), node-internal
config that never crosses .spec. No .spec consumer pins the old value: the
float mentions in astral-py and astral-js are primitive-codec vectors, not
op schemas. The one ecosystem type carrying floats, lna.channel_state_stats,
documents them as float64 — the width this change now publishes.

Restores the f64 row held back from TestFieldEditor_ObjectType and adds an
OpSpec-level guard, since the published spec is what the defect corrupted.
Both fail against the old mapping.

+++ lib/query/field_editor.go
@@ -59,7 +59,7 @@ func (editor FieldEditor) ObjectType() string {
 	case reflect.Float32:
 		return astral.Float32(0).ObjectType()
 	case reflect.Float64:
-		return astral.Float32(0).ObjectType()
+		return astral.Float64(0).ObjectType()
 	case reflect.Bool:
 		return astral.Bool(false).ObjectType()
